flushing out RBAC test for controllers --
Testing of controllers for role based access.  Implementing positive and negative
access tests.  Updated fixtures with additional users and roles to facilitate testing.
adjusted a couple of rubocop metrics.  Cucumber tests currently set the users role to
that of admin, in the future cucumber tests should be adjusted for user/roles.