Add product filtering for measure selection (#756)
* Convert to unix line endings

* Add filtering for custom measure selection

Adds the ability for users to filter when they choose to select custom measures. Also adds associated tests for the new functionality.

* Fix eslint errors

* Disable clear X in input box on IE

In Internet Explorer there is a clear button that is automatically added to all input fields. In this case it breaks since there is no event you can bind to that happens when that clear X is clicked. Disabling it is the most simplistic option in this case.

* Fix an intermittently failing test

This test seems to fail sometimes when find_by is called before the product is actually finished creating. By waiting until we are able to actually navigate to the product through the UI we are able to mitigate this issue.

* Bump max wait time for elements to appear for capybara