Fix uploading when behind an web server such as nginx
Currently when you attempt to upload a new bundle to the application when behind a proxy server such as nginx the upload fails. This is due to the fact that the admin_bundles_url returns the absolute url which is incorrect when behind a proxy server. If we use admin_bundles_path instead then it returns a relative url which does not encounter this problem.