Cleanup fakefs usage to be more localized and add app/ folder to fix autoloading
Strange issues were happening with fakefs not properly creating the cypress.yml file and causing tests to fail. This reworks the way it works so that it will only call it when we absolutely need it. Also we add the app/ directory so that autoloading works again. There were some cases where you could run the test suite with specific seeds and rails would try to autoload a function while in the FakeFS wrapper and would not be able to find the needed file. Adding the app/ directory fixes these cases.