Add a pending state in order to fix a race condition
Often when the test suite ran it would fail because the page was not refreshing properly due to a race condition between the state of the test and some other calls in the code. It was happening because the should_reload_product_test_link? function was being used in conjunction with the task_status to determine the test state. These would get out of sync and cause an incorrect code path to be taken, which would cause the page to stop refreshing even though the tests were not completed. This cleans up that code to make less queries which will cause the state to be more reliable.