Disable basic auth when users are not accessing via the JSON API
This is a bit hacky however it considerably helps app usability when dealing with the JSON API. Once a user authenticates with basic auth the browser keeps sending the credentials every time. By clearing the HTTP_AUTHORIZATION variable, unless the user is accessing via the JSON API we stop the app from allowing basic auth for the regular non-JSON parts of the application.