Correct issue where measure filtering was not working when only 1 bundle was imported
We were assuming that when there is only 1 bundle imported that we would still be able to grab the currently checked bundle to figure out which measures are currently available, however this is not correct. We can instead try to grab any checked bundles and if none are checked assume that there is only 1 and then grab it instead.