Restructure Patient model for easier integration of Bring your own Patient
This adds 3 subclasses of Patient, BundlePatient, VendorPatient, and ProductTestPatient
These 3 subclasses allow us to take advantage of Rails Single Table Inheritance, providing us with a :_type field on the Patient model. This type field allows us to properly use has_many relations in the app again which was previously not possible on models such as the bundle model.
Additionally, this has restructured the bundle import to use insert_many instead of insert_one which provides a huge performance speedup during bundle import.